refactor: RecommendUiState 패키지 변경
